+++ backend/src/app/core/data/crud/project.py
@@ -88,16 +88,9 @@ class CRUDProject(CRUDBase[ProjectORM, ProjectCreate, ProjectUpdate]):
 
         return user_db_obj
 
-    def exists_by_user_and_title(self, db: Session, user_id: int, title: str) -> bool:
+    def exists_by_title(self, db: Session, title: str) -> bool:
         return (
-            db.query(self.model)
-            .join(
-                UserORM,
-                UserORM.id == user_id,
-            )
-            .filter(self.model.title == title)
-            .first()
-            is not None
+            db.query(self.model).filter(self.model.title == title).first() is not None
         )
