+++ backend/src/app/core/data/dto/import_job.py
@@ -9,23 +9,9 @@ from app.core.data.dto.dto_base import UpdateDTOBase
 
 
 class ImportJobType(str, Enum):
-    SINGLE_PROJECT_ALL_DATA = "SINGLE_PROJECT_ALL_DATA"
-    SINGLE_PROJECT_ALL_USER = "SINGLE_PROJECT_ALL_USER"
-    SINGLE_PROJECT_ALL_TAGS = "SINGLE_PROJECT_ALL_TAGS"
-    SINGLE_PROJECT_SELECTED_SDOCS = "SINGLE_PROJECT_SELECTED_SDOCS"
-
-    SINGLE_USER_ALL_DATA = "SINGLE_USER_ALL_DATA"
-    SINGLE_PROJECT_ALL_METADATA = "SINGLE_PROJECT_ALL_METADATA"
-    SINGLE_PROJECT_ALL_PROJECT_PROJECT_METADATA = (
-        "SINGLE_PROJECT_ALL_PROJECT_PROJECT_METADATA"
-    )
-    SINGLE_PROJECT = "SINGLE_PROJECT"
-    SINGLE_USER_ALL_CODES = "SINGLE_USER_ALL_CODES"
-    SINGLE_USER_ALL_MEMOS = "SINGLE_USER_ALL_MEMOS"
-    SINGLE_USER_LOGBOOK = "SINGLE_USER_LOGBOOK"
-
-    SINGLE_DOC_ALL_USER_ANNOTATIONS = "SINGLE_DOC_ALL_USER_ANNOTATIONS"
-    SINGLE_DOC_SINGLE_USER_ANNOTATIONS = "SINGLE_DOC_SINGLE_USER_ANNOTATIONS"
+    PROJECT = "PROJECT"
+    CODES = "CODES"
+    TAGS = "TAGS"
 
 
 class ImportJobParameters(BaseModel):
